requirements.txt
demucs
ffmpeg-python
webrtcvad
openai-whisper
praat-parselmouth
